AsyncMessageSequence clean up, made optional decoding parameter types Sendable.
