+++ Sources/SwiftProtobuf/BinaryDecodingOptions.swift
@@ -13,7 +13,7 @@
 // -----------------------------------------------------------------------------
 
 /// Options for binary decoding.
-public struct BinaryDecodingOptions {
+public struct BinaryDecodingOptions: Sendable {
   /// The maximum nesting of message with messages.  The default is 100.
   ///
   /// To prevent corrupt or malicious messages from causing stack overflows,
