+++ Sources/SwiftProtobuf/ExtensionMap.swift
@@ -22,7 +22,7 @@
 /// This is a protocol so that developers can build their own
 /// extension handling if they need something more complex than the
 /// standard `SimpleExtensionMap` implementation.
-public protocol ExtensionMap {
+public protocol ExtensionMap: Sendable {
     /// Returns the extension object describing an extension or nil
     subscript(messageType: Message.Type, fieldNumber: Int) -> AnyMessageExtension? { get }
